Extend StrictOptimizedSeqOps optimizations to mutable collections
